Vastly simplify and refactor the tasks
This mainly includes the complete removal of all the logic involved with
storing and deleting multiple versions of teamspeak, which is pointless.

All the backup logic has been removed as well, because it is both easier
and simpler to just rsync (or the ansible equivalent) the hole ts3
folder over to the target machine.